+++ Laba/practica2/src/utils.cpp
@@ -1,33 +1,30 @@
-#include "utils.h"
+#include "../include/utils.h"
 #include <iostream>
 #include <fstream>
 #include <limits>
 #include <string>
+
 using namespace std;
 
-// Функция полной очистки буфера при ошибках ввода
 void clear_input() {
-    cin.clear(); // Сбрасываем флаг ошибки потока
-    cin.ignore(numeric_limits<streamsize>::max(), '\n'); // Очищаем всё до конца строки
+    cin.clear(); 
+    cin.ignore(numeric_limits<streamsize>::max(), '\n'); 
 }
 
-uint get_ll(const string& podskazka, uint minZnachenie, uint maxZnachenie) {
-    uint znachenie;
+uint64_t get_ll(const string& podskazka, uint64_t minZnachenie, uint64_t maxZnachenie) {
+    uint64_t znachenie;
     while (true) {
         cout << podskazka;
         if (!(cin >> znachenie)) {
             cout << "Ошибка: введено не число. Попробуйте снова.\n";
-            clear_input(); // Здесь очистка обязательна, так как ввод сломался
+            clear_input(); 
             continue;
         }
         
         if (znachenie < minZnachenie || znachenie > maxZnachenie) {
             cout << "Ошибка: число должно быть в диапазоне [" 
                  << minZnachenie << ", " << maxZnachenie << "].\n";
-            // Если число считалось, но не подошло по диапазону, 
-            // буфер чистить не нужно, просто идем на новую итерацию
         } else {
-            // Перед успешным возвратом избавляемся от остатков символа '\n' в буфере,
             string pustishka;
             getline(cin, pustishka); 
             return znachenie;
@@ -40,7 +37,6 @@ string get_str(const string& podskazka) {
     cout << podskazka;
     cin >> s;
     
-
     string pustishka;
     getline(cin, pustishka);
     return s;
